% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balances.R
\name{compute_node_balances}
\alias{compute_node_balances}
\title{Computes balances (i.e., isometric log ratios, for each sample separately) of feature abundances at each node in the tree.}
\usage{
compute_node_balances(
  tree,
  abun_table,
  min_num_tips = 5,
  ncores = 1,
  pseudocount = NULL,
  subset_to_test = NULL
)
}
\arguments{
\item{tree}{Phylo object with tip labels matching row names of input abundance table. Note that node labels are required.}

\item{abun_table}{Abundance table, e.g., read counts or relative abundance.
Should be dataframe with column names correspond to sample names and row names corresponding to the tips of the tree.
No 0 values are permitted unless the "pseudocount" option is set.}

\item{min_num_tips}{Minimum number of tips that must be found on each side of a node for it to be included.}

\item{ncores}{Number of cores to use for steps of function that can be run in parallel.}

\item{pseudocount}{Constant to add to all abundance values, to ensure that there are only non-zero values. For read count data this would typically be 1.}

\item{subset_to_test}{Vector of node labels (\emph{not indices}) that correspond to the subset of nodes that should be considered.
Note that balances will still only be computed at each of these nodes if they have a sufficient number of underlying tips (as specified by the "min_num_tips" argument).}
}
\value{
List containing three items:

"tips_underlying_nodes": the tips on the left-hand side (lhs; the numerator) and right-hand side (rhs; the denominator) of each node.
Note that which side of the node is denoted as the left-hand or right-hand side is arbitrary.

"balances": list with each non-negligible nodes as a separate element. The sample balances for each node are provided as a numeric vector within each of these elements.

"negligible_nodes": character vector of node labels where there are fewer tips on either side of the node than specified by the "min_num_tips" argument.
}
\description{
Computes balances (i.e., isometric log ratios, for each sample separately) of feature abundances at each node in the tree.
}
