% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenize_modified.R
\name{fit.beta.list}
\alias{fit.beta.list}
\title{Get a distribution of environment prevalences from a matrix.
This function was taken unchanged from the phylogenize R codebase.}
\usage{
fit.beta.list(mtx, ids, fallback = c(NA, NA))
}
\arguments{
\item{mtx}{A matrix of presence/absence values.}

\item{ids}{A named factor assigning samples (matrix columns) to environments.}

\item{fallback}{A two-element numeric vector, giving the beta parameters to
use if fitting fails.}
}
\value{
A best-fit of prevalences to a beta distribution.
}
\description{
This is used to optimize the value of the free parameter $b$ in
\code{regularize.pET}.
}
\keyword{internal}
