% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenize_modified.R
\name{logistic}
\alias{logistic}
\title{Calculate inverse-logit of a value or vector of values.
This function was taken unchanged from the phylogenize R codebase.}
\usage{
logistic(x)
}
\arguments{
\item{x}{Numeric value, or numeric vector of numeric values.}
}
\description{
Calculate inverse-logit of a value or vector of values.
This function was taken unchanged from the phylogenize R codebase.
}
