% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenize_modified.R
\name{logit}
\alias{logit}
\title{Calculate logit of a value or vector of values.
This function was taken unchanged from the phylogenize R codebase
(except it is no longer exported).}
\usage{
logit(x)
}
\description{
Calculate logit of a value or vector of values.
This function was taken unchanged from the phylogenize R codebase
(except it is no longer exported).
}
