% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenize_modified.R
\name{optimize_b_wrapper}
\alias{optimize_b_wrapper}
\title{Based on a real presence/absence matrix, optimize the value of the
regularization parameter $b$.}
\usage{
optimize_b_wrapper(
  real_abun_table,
  real_sample_values,
  focal_var_level,
  effect.size = 2,
  prior,
  tol,
  bounds = c(0.01, 5),
  add.pc = TRUE,
  a = 0.05,
  pos.prop = 0.5
)
}
\description{
Based on a real presence/absence matrix, optimize the value of the
regularization parameter $b$.
}
