% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{POMS_pipeline}
\alias{POMS_pipeline}
\title{Two-group POMS pipeline}
\usage{
POMS_pipeline(
  abun,
  func,
  phylogeny,
  group1_samples,
  group2_samples,
  ncores = 1,
  pseudocount = 1,
  significant_nodes = NULL,
  tested_balances = NULL,
  min_num_tips = 10,
  min_func_instances = 10,
  min_func_prop = 0.001,
  multinomial_min_sig = 5,
  balance_p_cutoff = 0.05,
  balance_correction = "none",
  function_p_cutoff = 0.05,
  function_correction = "none",
  func_descrip_infile = NULL,
  run_multinomial_test = TRUE,
  multinomial_correction = "BH",
  calc_node_dist = FALSE,
  detailed_output = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{abun}{Dataframe of MAG or ASV abundances}
}
\value{
A matrix of the infile
}
\description{
This function will identify significant nodes based on Wilcoxon tests
and then identify enriched functions.
}
