% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenize_modified.R
\name{regularize.pET}
\alias{regularize.pET}
\title{Obtain a regularized estimate of specificity.
This function was taken unchanged from the phylogenize R codebase.}
\usage{
regularize.pET(
  vec,
  env.ids,
  which.env = 1,
  prior = 0.05,
  b = 1,
  add.pc = FALSE,
  min.limit = -10,
  max.limit = 10
)
}
\arguments{
\item{vec}{A named numeric vector giving presence/absence across samples.}

\item{env.ids}{A named factor assigning an environment to each sample (names).}

\item{which.env}{A string: in which environment should specificity be calculated?}

\item{prior}{Prior probability of \code{which.env}}

\item{b}{Free parameter giving degree of regularization (see
\code{optimize.b.wrapper}).}

\item{add.pc}{Boolean giving whether to add a pseudocount.}

\item{min.limit}{Will not optimize below this value.}

\item{max.limit}{Will not optimize above this value.}
}
\value{
A list. \code{x}: regularized specificity estimate; \code{p}:
regularized prevalence estimate; \code{x.init}: naive specificity
estimate; \code{p.init}: naive prevalence estimate; \code{pT}:
probability of encountering a particular taxon, marginalized across
environments
}
\description{
Obtain a regularized estimate of specificity.
This function was taken unchanged from the phylogenize R codebase.
}
\keyword{internal}
