% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenize_modified.R
\name{score.regularization}
\alias{score.regularization}
\title{Score a simulated regularization by how well it recapitulates the ground truth.
This function was taken unchanged from the phylogenize R codebase.}
\usage{
score.regularization(
  mtx,
  ids,
  real.fx,
  which.env = 2,
  prior = 0.002,
  b = 0.1,
  add.pc = FALSE,
  tol = prior * 0.005,
  ...
)
}
\description{
Score a simulated regularization by how well it recapitulates the ground truth.
This function was taken unchanged from the phylogenize R codebase.
}
