% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{two_group_balance_tree_pipeline}
\alias{two_group_balance_tree_pipeline}
\title{Two-group POMS pipeline}
\usage{
two_group_balance_tree_pipeline(
  abun,
  func,
  phylogeny,
  taxa,
  group1_samples,
  group2_samples,
  ncores = 1,
  pseudocount = 1,
  significant_nodes = NULL,
  tested_balances = NULL,
  min_num_tips = 10,
  min_func_instances = 10,
  min_func_prop = 0.001,
  name_threshold = 0.9,
  balance_p_cutoff = 0.05,
  balance_correction = "BY",
  function_p_cutoff = 0.05,
  function_correction = "none",
 
    func_descrip_infile = "/home/gavin/projects/POMS/KEGG_mappings/prepped/KO_descrip_22Aug2019.tsv",
  verbose = FALSE
)
}
\arguments{
\item{abun}{Dataframe of MAG or ASV abundances}
}
\value{
A matrix of the infile
}
\description{
This function will identify significant nodes based on Wilcoxon tests
and then identify enriched functions.
}
