% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenize_modified.R
\name{lapply.across.names}
\alias{lapply.across.names}
\title{Apply a function to a vector of names, with the returned list having those
names.
This function was taken unchanged from the phylogenize R codebase.}
\usage{
lapply.across.names(X, FUN, ...)
}
\description{
Apply a function to a vector of names, with the returned list having those
names.
This function was taken unchanged from the phylogenize R codebase.
}
