% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenize_modified.R
\name{logit}
\alias{logit}
\title{Calculate logit of a value or vector of values.
This function was taken unchanged from the phylogenize R codebase.}
\usage{
logit(x)
}
\arguments{
\item{x}{Numeric value, or numeric vector of numeric values.}
}
\description{
Calculate logit of a value or vector of values.
This function was taken unchanged from the phylogenize R codebase.
}
