% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenize_modified.R
\name{regularize.pET}
\alias{regularize.pET}
\title{Obtain a regularized estimate of specificity.
This function was taken unchanged from the phylogenize R codebase.}
\usage{
regularize.pET(
  vec,
  env.ids,
  which.env = 1,
  prior = 0.05,
  b = 1,
  add.pc = FALSE,
  min.limit = -10,
  max.limit = 10
)
}
\description{
Obtain a regularized estimate of specificity.
This function was taken unchanged from the phylogenize R codebase.
}
